[nits] rename librabft-v2/src/base_types.rs
